+++ src/main/java/de/maxhenkel/voicechat/voice/client/AudioChannel.java
@@ -68,6 +68,14 @@ public class AudioChannel extends Thread {
             speaker.open(af);
             gainControl = (FloatControl) speaker.getControl(FloatControl.Type.MASTER_GAIN);
             while (!stopped) {
+
+                if (VoicechatClient.CLIENT.getPlayerStateManager().isDisabled()) {
+                    speaker.stop();
+                    queue.clear();
+                    closeAndKill();
+                    return;
+                }
+
                 // Stopping the data line when the buffer is empty
                 // to prevent the last sound getting repeated
                 if (speaker.getBufferSize() - speaker.available() <= 0 && speaker.isActive()) {
@@ -107,7 +115,7 @@ public class AudioChannel extends Thread {
 
                 gainControl.setValue(Math.min(Math.max(Utils.percentageToDB(percentage * VoicechatClient.CLIENT_CONFIG.voiceChatVolume.get().floatValue() * (float) VoicechatClient.VOLUME_CONFIG.getVolume(player)), gainControl.getMinimum()), gainControl.getMaximum()));
 
-                Pair<Float, Float> stereoVolume = Utils.getStereoVolume(minecraft.player.getPos(), minecraft.player.yaw, player.getPos(), client.getVoiceChatDistance());
+                Pair<Float, Float> stereoVolume = Utils.getStereoVolume(minecraft, player.getPos(), client.getVoiceChatDistance());
 
                 byte[] stereo = Utils.convertToStereo(packet.getData(), stereoVolume.getLeft(), stereoVolume.getRight());
                 speaker.write(stereo, 0, stereo.length);