+++ changelog.md
@@ -1,1 +1,2 @@
-- Performance improvements
+- Fixed player icon rendering
+- Fixed sound coming from the wrong side when in F5